init guardrails on proxy
